Move conductor to lib and create common app.
