+++ .gitignore
@@ -129,11 +129,8 @@ dist/
 downloads/
 eggs/
 .eggs/
-lib/
-lib64/
 parts/
 sdist/
-var/
 *.egg-info/
 .installed.cfg
 *.egg
