+++ container/cli.py
@@ -12,7 +12,7 @@ import requests.exceptions
 
 from . import core
 from . import config
-from . import exceptions
+from .common import exceptions
 from .utils import load_shipit_engine, AVAILABLE_SHIPIT_ENGINES
 from .conductor import exceptions as conductor_exc
