+++ container/common/docker.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+from __future__ import absolute_import
+
+import logging
+
+logger = logging.getLogger(__name__)
+
