+++ container/config.py
@@ -9,9 +9,9 @@ import json
 import six
 
 from collections import Mapping
-from .exceptions import AnsibleContainerConfigException
-
 from ruamel import yaml, ordereddict
+
+from .common.exceptions import AnsibleContainerConfigException
 
 # TODO: Actually do some schema validation
