+++ container/lib/conductor/docker/files/get-pip.py
