+++ container/lib/conductor/docker/templates/conductor-dockerfile.j2
