+++ container/lib/conductor/templates/role/.travis.j2.yml
