+++ container/lib/conductor/templates/role/defaults/main.j2.yml
