+++ container/lib/conductor/templates/role/meta/main.j2.yml
