+++ test/unit/container/test_utils.py
@@ -1,11 +1,13 @@
+import os
 import shutil
 import tempfile
-from os import path
 import unittest
-import os
+from os import path
+
 import pytest
+
+from container.common.exceptions import AnsibleContainerNotInitializedException
 from container.utils import assert_initialized
-from container.exceptions import AnsibleContainerNotInitializedException
 
 
 class TestMissingFiles(unittest.TestCase):
